spot/requirements.txt
requests
PyPDF2
openpyxl
python-docx
ollama
langchain
langchain-community
langchain-core